Extract setBasicMethods
